backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
sqlparse==0.5.0
tzdata==2024.1
google-generativeai
python-dotenv
opencv-python
numpy
scipy
Pillow
transformers
whisper-timestamped
moviepy
torch